test: dedupe get_staging_metadata patch decorators via shared fixture
Moved the repeated six @patch decorators into a staging_metadata_deps
fixture in a new conftest.py, shared by test_staging_database.py and
test_staging_api.py.
