+++ apps/backend/tests/api/routes/test_staging_database.py
@@ -1,6 +1,5 @@
 """Tests for database source type in staging endpoints."""
 
-from collections.abc import Generator
 from datetime import datetime, timezone
 from types import SimpleNamespace
 from unittest.mock import MagicMock, patch
@@ -218,24 +217,6 @@ class TestDagSuccessCallbackDeleteGuard:
         mock_delete.assert_called_once_with("/tmp/somefile.csv")
 
 
-@pytest.fixture
-def staging_metadata_deps() -> Generator[SimpleNamespace, None, None]:
-    """Patch every dependency get_staging_metadata needs besides the IntegrityLink
-    and staging table content, which each test configures itself via .load and
-    .table."""
-    with (
-        patch("src.api.routes.ingestion.staging.get_staging_schema", return_value="staging"),
-        patch("src.api.routes.ingestion.staging.select"),
-        patch("src.api.routes.ingestion.staging.Table") as mock_table,
-        patch("src.api.routes.ingestion.staging._resolve_columns") as mock_resolve_cols,
-        patch("src.api.routes.ingestion.staging._detect_original_projection") as mock_detect_proj,
-        patch("src.api.routes.ingestion.staging.load_authorized_integrity_link") as mock_load,
-    ):
-        mock_resolve_cols.return_value = ([], None)
-        mock_detect_proj.return_value = None
-        yield SimpleNamespace(load=mock_load, table=mock_table)
-
-
 def _call_get_staging_metadata() -> StagingMetadataResponse:
     data_session = MagicMock()
     data_session.scalar.return_value = 0
@@ -251,20 +232,8 @@ def _call_get_staging_metadata() -> StagingMetadataResponse:
 class TestGetStagingMetadataTitleFallback:
     """Test title fallback logic in get_staging_metadata for database sources."""
 
-    @patch("src.api.routes.ingestion.staging.get_staging_schema", return_value="staging")
-    @patch("src.api.routes.ingestion.staging.select")
-    @patch("src.api.routes.ingestion.staging.Table")
-    @patch("src.api.routes.ingestion.staging._resolve_columns")
-    @patch("src.api.routes.ingestion.staging._detect_original_projection")
-    @patch("src.api.routes.ingestion.staging.load_authorized_integrity_link")
     def test_title_falls_back_to_table_name_from_source_url(
-        self,
-        mock_load: MagicMock,
-        mock_detect_proj: MagicMock,
-        mock_resolve_cols: MagicMock,
-        mock_table: MagicMock,
-        mock_select: MagicMock,
-        mock_get_schema: MagicMock,
+        self, staging_metadata_deps: SimpleNamespace
     ) -> None:
         """Title is the table name parsed from db://{schema}/{table} when no custom title is set."""
         mock_link = MagicMock()
@@ -275,37 +244,14 @@ class TestGetStagingMetadataTitleFallback:
         mock_link.source_url = "db://SOURCE_DB_1/geo/parcels"
         mock_link.integrity_transformation = None
         mock_link.final_table_name = None
-        mock_load.return_value = (mock_link, MagicMock())
-        mock_resolve_cols.return_value = ([], None)
-        mock_detect_proj.return_value = None
+        staging_metadata_deps.load.return_value = (mock_link, MagicMock())
 
-        data_session = MagicMock()
-        data_session.scalar.return_value = 0
-
-        result = get_staging_metadata(
-            data_session=data_session,
-            datafeeder_session=MagicMock(),
-            geo_ctx=MagicMock(),
-            integrity_link_id=str(uuid4()),
-            group_ids=[],
-        )
+        result = _call_get_staging_metadata()
 
         assert result.title == "parcels"
 
-    @patch("src.api.routes.ingestion.staging.get_staging_schema", return_value="staging")
-    @patch("src.api.routes.ingestion.staging.select")
-    @patch("src.api.routes.ingestion.staging.Table")
-    @patch("src.api.routes.ingestion.staging._resolve_columns")
-    @patch("src.api.routes.ingestion.staging._detect_original_projection")
-    @patch("src.api.routes.ingestion.staging.load_authorized_integrity_link")
     def test_custom_title_overrides_table_name(
-        self,
-        mock_load: MagicMock,
-        mock_detect_proj: MagicMock,
-        mock_resolve_cols: MagicMock,
-        mock_table: MagicMock,
-        mock_select: MagicMock,
-        mock_get_schema: MagicMock,
+        self, staging_metadata_deps: SimpleNamespace
     ) -> None:
         """integrity_title takes precedence over table name derived from source_url."""
         mock_link = MagicMock()
@@ -316,37 +262,14 @@ class TestGetStagingMetadataTitleFallback:
         mock_link.source_url = "db://SOURCE_DB_1/geo/parcels"
         mock_link.integrity_transformation = None
         mock_link.final_table_name = None
-        mock_load.return_value = (mock_link, MagicMock())
-        mock_resolve_cols.return_value = ([], None)
-        mock_detect_proj.return_value = None
+        staging_metadata_deps.load.return_value = (mock_link, MagicMock())
 
-        data_session = MagicMock()
-        data_session.scalar.return_value = 0
-
-        result = get_staging_metadata(
-            data_session=data_session,
-            datafeeder_session=MagicMock(),
-            geo_ctx=MagicMock(),
-            integrity_link_id=str(uuid4()),
-            group_ids=[],
-        )
+        result = _call_get_staging_metadata()
 
         assert result.title == "Parcelles cadastrales"
 
-    @patch("src.api.routes.ingestion.staging.get_staging_schema", return_value="staging")
-    @patch("src.api.routes.ingestion.staging.select")
-    @patch("src.api.routes.ingestion.staging.Table")
-    @patch("src.api.routes.ingestion.staging._resolve_columns")
-    @patch("src.api.routes.ingestion.staging._detect_original_projection")
-    @patch("src.api.routes.ingestion.staging.load_authorized_integrity_link")
     def test_title_strips_extension_from_source_file_name(
-        self,
-        mock_load: MagicMock,
-        mock_detect_proj: MagicMock,
-        mock_resolve_cols: MagicMock,
-        mock_table: MagicMock,
-        mock_select: MagicMock,
-        mock_get_schema: MagicMock,
+        self, staging_metadata_deps: SimpleNamespace
     ) -> None:
         """source_file_name extension is stripped when used as title fallback."""
         mock_link = MagicMock()
@@ -357,38 +280,13 @@ class TestGetStagingMetadataTitleFallback:
         mock_link.source_url = None
         mock_link.integrity_transformation = None
         mock_link.final_table_name = None
-        mock_load.return_value = (mock_link, MagicMock())
-        mock_resolve_cols.return_value = ([], None)
-        mock_detect_proj.return_value = None
+        staging_metadata_deps.load.return_value = (mock_link, MagicMock())
 
-        data_session = MagicMock()
-        data_session.scalar.return_value = 0
-
-        result = get_staging_metadata(
-            data_session=data_session,
-            datafeeder_session=MagicMock(),
-            geo_ctx=MagicMock(),
-            integrity_link_id=str(uuid4()),
-            group_ids=[],
-        )
+        result = _call_get_staging_metadata()
 
         assert result.title == "data"
 
-    @patch("src.api.routes.ingestion.staging.get_staging_schema", return_value="staging")
-    @patch("src.api.routes.ingestion.staging.select")
-    @patch("src.api.routes.ingestion.staging.Table")
-    @patch("src.api.routes.ingestion.staging._resolve_columns")
-    @patch("src.api.routes.ingestion.staging._detect_original_projection")
-    @patch("src.api.routes.ingestion.staging.load_authorized_integrity_link")
-    def test_title_preserves_hidden_file_name(
-        self,
-        mock_load: MagicMock,
-        mock_detect_proj: MagicMock,
-        mock_resolve_cols: MagicMock,
-        mock_table: MagicMock,
-        mock_select: MagicMock,
-        mock_get_schema: MagicMock,
-    ) -> None:
+    def test_title_preserves_hidden_file_name(self, staging_metadata_deps: SimpleNamespace) -> None:
         """Hidden-file names (leading dot) are kept as-is without stripping the extension."""
         mock_link = MagicMock()
         mock_link.integrity_title = None
@@ -398,20 +296,9 @@ class TestGetStagingMetadataTitleFallback:
         mock_link.source_url = None
         mock_link.integrity_transformation = None
         mock_link.final_table_name = None
-        mock_load.return_value = (mock_link, MagicMock())
-        mock_resolve_cols.return_value = ([], None)
-        mock_detect_proj.return_value = None
+        staging_metadata_deps.load.return_value = (mock_link, MagicMock())
 
-        data_session = MagicMock()
-        data_session.scalar.return_value = 0
-
-        result = get_staging_metadata(
-            data_session=data_session,
-            datafeeder_session=MagicMock(),
-            geo_ctx=MagicMock(),
-            integrity_link_id=str(uuid4()),
-            group_ids=[],
-        )
+        result = _call_get_staging_metadata()
 
         assert result.title == ".hidden"
